started a firstrun script to setup firefox et. al

+++ salt/pacpub/files/autostart_kiosk.sh
@@ -1,6 +1,7 @@
 #!/usr/bin/env bash
 #usr/local/bin/autostart_kiosk.sh - run as root
-# force run dhclient to update network + hostname from dhcp and dns
-#dhclient eth0
 # restore kioskuser upon login
-rsync -qa --delete --exclude='.Xauthority' --exclude='.cache' /opt/kiosk/ $HOME/
+if [ -d "/opt/kiosk" ]
+then
+  rsync -qa --delete --exclude='.Xauthority' --exclude='.cache' /opt/kiosk/ $HOME/
+fi
